refactored config script template files into separate directory
